move logging to actor
