+++ backpack/CS455-555-1-s20-team/p1-chatserver-filesync/src/main/scala/ActorStuff.scala
@@ -1,6 +1,11 @@
-import akka.actor.typed.{ ActorRef, ActorSystem, Behavior }
-import akka.actor.Actor;
-import os._;
+import java.io.{FileWriter, PrintWriter}
+import java.nio.file.Path
+import java.text.SimpleDateFormat
+import java.util.{Calendar, Locale, TimeZone}
+
+import akka.actor.typed.{ActorRef, ActorSystem, Behavior}
+import akka.actor.Actor
+//import os._;
 
 abstract sealed class IOCmd
 case class Write(s: String)
@@ -13,6 +18,14 @@ final class IOActor(outfile: Path) extends Actor {
   }
 
   private[this] def write(s: Seq[String]) = {
-
+    val timeZone = TimeZone.getTimeZone("UTC")
+    val dateFormat: SimpleDateFormat = new SimpleDateFormat("EE MMM dd HH:mm:ss zzz yyyy", Locale.US)
+    val dateTime = Calendar.getInstance(timeZone)
+    dateFormat.setTimeZone(timeZone)
+    val currentHour = dateFormat.format(dateTime.getTime)
+    val fileWriter = new FileWriter(outfile.toFile,true)
+    val printWriter = new PrintWriter(fileWriter)
+    printWriter.println(s"$currentHour ${s.mkString("\n")}")
+    printWriter.close()
   }
 }
